# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight REST relay for managing TR-069 devices via the GenieACS NBI. v2.2.0 adds 25 new operational endpoints across 4 phases (CPE lifecycle, NOC support, customer self-service, GenieACS metadata) on top of the v2.1.0 baseline. 100% main-package coverage, 40/40 endpoints end-to-end verified on real ZTE F670L V9.0.10P1N12A via VPN lab (sessions 5i + 5j 2026-04-15).\nSee CHANGELOG.md [2.2.0] for full release notes and upstream genieacs-stack v1.3.1 blocker for customer factory-reset workflows.",
        "title": "GenieACS Relay API",
        "contact": {
            "name": "Cepat Kilat Teknologi",
            "url": "https://github.com/Cepat-Kilat-Teknologi/genieacs-relay"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/genieacs",
    "paths": {
        "/admin-password/{ip}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the CPE local web admin password via TR-069 SetParameterValues on `InternetGatewayDevice.UserInterface.WebPassword`. Distinct from PPPoE credentials and TR-069 ACS auth. Password is NOT echoed in the response or audit logs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Set CPE admin web password",
                "parameters": [
                    {
                        "type": "string",
                        "example": "192.168.1.1",
                        "description": "Device IP address",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New admin password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SetAdminPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.SetAdminPasswordResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
...
        "/ntp/{ip}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the CPE NTP server list (max 5 entries) and/or timezone via TR-069 SetParameterValues. Standard TR-098 paths: `InternetGatewayDevice.Time.NTPServer{1..5}` and `InternetGatewayDevice.Time.LocalTimeZoneName`. At least one of ntp_servers or timezone must be provided.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Set CPE NTP servers and timezone",
                "parameters": [
                    {
                        "type": "string",
                        "example": "192.168.1.1",
                        "description": "Device IP address",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "NTP / timezone payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SetNTPRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.SetNTPResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "data": {
                    "description": "Response payload (success) OR error detail (string/field list)"
                },
                "error_code": {
                    "description": "Machine-readable error code (VALIDATION_ERROR, NOT_FOUND, ...)",
                    "type": "string"
                },
                "request_id": {
                    "description": "Correlation ID from X-Request-ID header",
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" for 2xx, HTTP reason phrase for errors",
                    "type": "string"
                }
            }
        },
...
        "main.SetAdminPasswordRequest": {
            "description": "Admin web password update request. The password is the new value to set on the CPE's local web admin interface — it does NOT change PPPoE credentials (use /pppoe/{ip} for that) or TR-069 ACS auth.",
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "main.SetAdminPasswordResponse": {
            "description": "Admin web password update response. Does NOT echo the password value back to keep it out of relay audit logs.",
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "main.SetNTPRequest": {
            "description": "NTP / timezone update request. At least one of ntp_servers or timezone must be provided.",
            "type": "object",
            "properties": {
                "ntp_servers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "main.SetNTPResponse": {
            "description": "NTP / timezone update response.",
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "ntp_servers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for authentication (required when MIDDLEWARE_AUTH=true)",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}